+++ R/cache.R
@@ -1647,13 +1647,13 @@ CacheDigest <- function(objsToDigest, ..., algo = "xxhash64", calledFrom = "Cach
     preDigest <- append(preDigest, preDigestQuick)
   }
 
-  res <- if (isTRUE(getOption("reproducible.useNewDigestAlgorithm") > 0)) {
+  # res <- if (isTRUE(getOption("reproducible.useNewDigestAlgorithm") > 0)) {
     .robustDigest(unname(sort(unlist(preDigest))), algo = algo, quick = TRUE, ...)
-  } else {
-    .requireNamespace("fastdigest", stopOnFALSE = TRUE,
-                      messageStart = "to use options('reproducible.useNewDigestAlgorithm' = FALSE")
-    fastdigest::fastdigest(preDigest)
-  }
+  # } else {
+  #   .requireNamespace("fastdigest", stopOnFALSE = TRUE,
+  #                     messageStart = "to use options('reproducible.useNewDigestAlgorithm' = FALSE")
+  #   fastdigest::fastdigest(preDigest)
+  # }
   list(outputHash = res, preDigest = preDigest)
 }