+++ R/options.R
@@ -210,7 +210,7 @@ reproducibleOptions <- function() {
     reproducible.useDBI = TRUE,
     # reproducible.useGDAL = FALSE, #
     reproducible.useMemoise = FALSE, #memoise
-    reproducible.useNewDigestAlgorithm = 2, # TODO: change in next release
+    # reproducible.useNewDigestAlgorithm = 2, # TODO: change in next release
     reproducible.useMultipleDBFiles = FALSE,
     reproducible.useTerra = TRUE, # Default for now
     # reproducible.useTerra = FALSE, # Default for now